done api CRUD subject co phan quyen

+++ src/main/java/com/doan/cnpm/controller/SubjectController.java
@@ -1,22 +1,23 @@
-package com.doan.cnpm.controller;
+    package com.doan.cnpm.controller;
 
 import com.doan.cnpm.domain.Subject;
-import com.doan.cnpm.domain.TutorDetails;
+import com.doan.cnpm.domain.User;
 import com.doan.cnpm.repositories.SubjectRepository;
+import com.doan.cnpm.repositories.UserRepository;
 import com.doan.cnpm.service.SubjectService;
 import com.doan.cnpm.service.dto.SubjectDTO;
-import com.doan.cnpm.service.dto.TutorDetailsDTO;
-import com.doan.cnpm.service.exception.TutorNotFoundException;
+import com.doan.cnpm.service.exception.AccessDeniedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
+import com.doan.cnpm.service.UserAuthorityService;
 
 import java.util.List;
+import java.util.Optional;
 
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 
 
@@ -27,44 +28,82 @@ public class SubjectController {
     private final Logger log = LoggerFactory.getLogger(SubjectController.class);
     private final SubjectRepository subjectRepository;
     private SubjectService subjectService;
+    private final UserRepository userRepository;
 
-    public SubjectController(SubjectRepository subjectRepository, SubjectService subjectService) {
+    private final UserAuthorityService userAuthorityService;
+
+    public SubjectController(SubjectRepository subjectRepository, SubjectService subjectService, UserRepository userRepository, UserAuthorityService userAuthorityService) {
         this.subjectRepository = subjectRepository;
         this.subjectService = subjectService;
+        this.userRepository = userRepository;
+        this.userAuthorityService = userAuthorityService;
     }
     
     @GetMapping(value="/v1/subject")
-    public List<Subject> getALlSubject(){
-        return subjectRepository.findAll();
+    public List<Subject> getALlSubject(HttpServletRequest request){
+        String username = request.getHeader("username");
+        Optional<User> user = userRepository.findOneByUsername(username);
+        String userId = String.valueOf(user.get().getId());
+        String authority = userAuthorityService.getAuthority(userId);
+        if(authority.equals("ROLE_ADMIN")) {
+            return subjectRepository.findAll();
+        }
+        throw new AccessDeniedException();
     }
 
     @GetMapping("v1/subject/details")
-    public ResponseEntity<Subject> getTutorDetails (HttpServletRequest request) {
+    public ResponseEntity<Subject> getTutorDetails (@RequestParam(name = "id") Long id, HttpServletRequest request) {
 
-        Long id = Long.parseLong( request.getHeader("id"));
-
-        Subject data = subjectRepository.findOneById(id);
-
-        return new ResponseEntity<>(data,HttpStatus.OK);
+        String username = request.getHeader("username");
+        Optional<User> user = userRepository.findOneByUsername(username);
+        String userId = String.valueOf(user.get().getId());
+        String authority = userAuthorityService.getAuthority(userId);
+        if(authority.equals("ROLE_ADMIN")) {
+            Subject data = subjectRepository.findOneById(id);
+            return new ResponseEntity<>(data, HttpStatus.OK);
+        }
+        throw new AccessDeniedException();
     }
 
     @PostMapping("v1/subject/create")
     @ResponseStatus(HttpStatus.CREATED)
     public void createSubject (@RequestBody SubjectDTO subject, HttpServletRequest request){
-        subjectService.CreateSubject(subject);
+        String username = request.getHeader("username");
+        Optional<User> user = userRepository.findOneByUsername(username);
+        String userId = String.valueOf(user.get().getId());
+        String authority = userAuthorityService.getAuthority(userId);
+        if(authority.equals("ROLE_ADMIN")) {
+            subjectService.CreateSubject(subject);
+            return;
+        }
+        throw new AccessDeniedException();
     }
 
     @PutMapping("v1/subject/update")
     @ResponseStatus(HttpStatus.UPGRADE_REQUIRED)
-    public void updateSubject (@RequestBody SubjectDTO tutor, HttpServletRequest request)  {
-        Long id = Long.parseLong( request.getHeader("id"));
-        subjectService.UpdateSubject(tutor, id);
+    public void updateSubject (@RequestBody SubjectDTO subject, HttpServletRequest request,@RequestParam(name = "id") Long id )  {
+        String username = request.getHeader("username");
+        Optional<User> user = userRepository.findOneByUsername(username);
+        String userId = String.valueOf(user.get().getId());
+        String authority = userAuthorityService.getAuthority(userId);
+        if(authority.equals("ROLE_ADMIN")) {
+            subjectService.UpdateSubject(subject, id);
+            return;
+        }
+        throw new AccessDeniedException();
     }
 
     @DeleteMapping("v1/subject/delete")
-    public void deleteSubject(HttpServletRequest request)
+    public void deleteSubject(HttpServletRequest request,@RequestParam(name = "id") Long id)
     {
-        Long id = Long.parseLong( request.getHeader("id"));
-        subjectService.DeleteSubject(id);
+        String username = request.getHeader("username");
+        Optional<User> user = userRepository.findOneByUsername(username);
+        String userId = String.valueOf(user.get().getId());
+        String authority = userAuthorityService.getAuthority(userId);
+        if(authority.equals("ROLE_ADMIN")) {
+            subjectService.DeleteSubject(id);
+            return;
+        }
+        throw new AccessDeniedException();
     }
 }
